Deprecate `Logger.Level.style` and `Logger.Level.name`. Fix a couple missing and typoed documentation comments on ConsoleLogger.
